requirements.txt
requests
json
webbbrowser
pprint
matplotlib.pyplot
sqlite3